+++ src/syncer/partner_web_client.ts
@@ -57,8 +57,14 @@ export class PartnerWebClient<
       this.incomingQueue.close();
     };
 
-    this.socket.onerror = (err) => {
-      console.error(err);
+    this.socket.onerror = (event) => {
+      if ("error" in event) {
+        this.incomingQueue.close({
+          withError: event.error,
+        });
+
+        return;
+      }
 
       this.incomingQueue.close({
         withError: new EarthstarError("Websocket error."),