+++ src/syncer/transfer_manager.ts
@@ -62,44 +62,6 @@ export class TransferManager<FormatsType, IncomingAttachmentSourceType> {
 
   // pass a syncagent's isDone to this
   registerSyncAgent(agent: SyncAgent<FormatsType>) {
-    // create a sealed thing for when all syncagents are done syncing docs.
-
-    const existingDocsStream = agent.replica.getQueryStream(
-      { orderBy: "localIndex ASC" },
-      "existing",
-      this.formats,
-    );
-
-    const { formatsLookup } = this;
-    const handleDownload = this.handleDownload.bind(this);
-
-    const pipedExistingDocsToManager = existingDocsStream.pipeTo(
-      new WritableStream<
-        QuerySourceEvent<FormatDocType<FormatsType>>
-      >({
-        async write(event) {
-          if (event.kind === "existing" || event.kind === "success") {
-            // Get the right format here...
-            const format = formatsLookup[event.doc.format];
-
-            const res = await agent.replica.getAttachment(event.doc, format);
-
-            if (isErr(res)) {
-              // This doc can't have an attachment attached. Do nothing.
-              return;
-            } else if (res === undefined) {
-              await handleDownload(
-                event.doc,
-                agent.replica,
-                agent.counterpartId,
-              );
-            }
-          }
-        },
-      }),
-    );
-
-    this.madeAllAttachmentRequestsEnroller.enrol(pipedExistingDocsToManager);
     this.madeAllAttachmentRequestsEnroller.enrol(agent.isDone());
   }
 
@@ -157,7 +119,6 @@ export class TransferManager<FormatsType, IncomingAttachmentSourceType> {
 
       promise.resolve();
 
-      // The sync agent will send a blob req.
       return "no_attachment";
     }