+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## NEXT
+
+- (Improvement) - Warn in the console when a replica could not ingest an
+  attachment during sync.
+- (Improvement) - Better error messages for web syncing failures (e.g. 404,
+  wrong endpoint).
+
 ## v10.0.0
 
 This is a major release which introduces attachments, share keypairs, efficient